баг фикс со спамом комманды настроек + ренейминг + начало разработки изменения оформления постов

+++ Classes/Others/Sender.cs
@@ -10,50 +10,51 @@ using System;
 
 namespace SendersNS {
     class Sender {
-        public static async Task<Message> SendTextMessage(long userId, string text) {
-            return await bot.SendTextMessageAsync(text: text, chatId: userId/*, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html*/);
+        private static string GeneralTextMessageStyling(string text) { return $"<b>{text}</b>"; }
+
+        public static async Task<Message> SendTextMessageAsync(long userId, string text) {
+            return await bot.SendTextMessageAsync(chatId: userId, text: GeneralTextMessageStyling(text), parseMode:Telegram.Bot.Types.Enums.ParseMode.Html );
         }
-        public static async Task<Message> SendTextMessage(long userId, string text, InlineKeyboardMarkup keyboard) {
-            return await bot.SendTextMessageAsync(text: text, chatId: userId, replyMarkup: keyboard/*, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html*/);
+        public static int SendTextMessageReturningMsgId(long userId, string text, InlineKeyboardMarkup keyboard) {
+            var msg = bot.SendTextMessageAsync(chatId: userId, text: GeneralTextMessageStyling(text), replyMarkup: keyboard, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
+            return msg.Result.MessageId;
+        }
+        public static async Task<Message> SendTextMessageAsync(long userId, string text, InlineKeyboardMarkup keyboard) {
+            return await bot.SendTextMessageAsync(chatId: userId, text: GeneralTextMessageStyling(text), replyMarkup: keyboard, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
         }
         public static async Task SendAutoDeleteMessage(long userId, string text) {
-            var autodelete_msg = await SendTextMessage(userId, text);
-            await Task.Delay(5000);
+            var autodelete_msg = await SendTextMessageAsync(userId, text);
+            await Task.Delay(10000);
             await bot.DeleteMessageAsync(chatId: userId, messageId: autodelete_msg.MessageId);
         }
         public static async Task SendAutoDeleteMessage(long userId, string text, int delay) {
-            var autodelete_msg = await SendTextMessage(userId, text);
+            var autodelete_msg = await SendTextMessageAsync(userId, text);
             await Task.Delay(delay);
             await bot.DeleteMessageAsync(chatId: userId, messageId: autodelete_msg.MessageId);
         }
-        public static async Task<Message> EditTextMessage (long userId, int msgId, string txt, InlineKeyboardMarkup keyboard) {
-
-            return await bot.EditMessageTextAsync(chatId: userId, messageId: msgId, text:$"<strong><i>{txt, -10}</i></strong>" , replyMarkup: keyboard, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html) ;
+        public static async Task<Message> EditTextMessageAsync (long userId, int msgId, string text, InlineKeyboardMarkup keyboard) {
+            return await bot.EditMessageTextAsync(chatId: userId, messageId: msgId, text:GeneralTextMessageStyling(text) , replyMarkup: keyboard, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html) ;
         }
+        
         public static async Task CreateVideoPost(Message msg) {
-            Random rndmNumber = new Random();
-            string videoPath = @$"../../../VideoBuffer/tiktok{msg.Chat.Id}{rndmNumber.Next(1000)}.mp4";
-            var videoUrl = LinkCorrector(msg.Text);
+            string videoPath = @$"../../../VideoBuffer/tiktok{msg.Chat.Id}{new Random().Next()}.mp4";
+            var videoUrl = msg.Text;
             var userId = msg.Chat.Id;
-            var videoIsDownloading_msg = await bot.SendTextMessageAsync(chatId: userId, text: "Видео загружается...");
-            try { await TTDownloader.Download(msg.Text, videoPath); } catch {
+            var videoIsDownloading_msg = await SendTextMessageAsync(userId, "Видео загружается...");
+            await TTDownloader.Download(msg.Text, videoPath);
+            if (System.IO.File.Exists(videoPath)) {
+                FileStream fileStream = new(videoPath, FileMode.Open, FileAccess.Read);
+                await bot.SendVideoAsync(chatId: userId, video: fileStream, replyMarkup: await InlineKeyboard.GetVideoSendingKeyboard(userId, videoUrl));
+                await fileStream.DisposeAsync();
+                await bot.DeleteMessageAsync(chatId: userId, messageId: videoIsDownloading_msg.MessageId);
+                await bot.DeleteMessageAsync(chatId: userId, messageId: msg.MessageId);
+                System.IO.File.Delete(videoPath);
+            } else {
                 await bot.DeleteMessageAsync(chatId: userId, messageId: videoIsDownloading_msg.MessageId);
                 await bot.DeleteMessageAsync(chatId: userId, messageId:msg.MessageId);
                 await SendAutoDeleteMessage(userId, "Видео недоступно");
-                Logs.GetErrorLogAboutTTDownloader(userId, videoUrl);
-                return;
+                await Logs.GetErrorLogAboutTTDownloaderAsync( videoUrl);
             }
-            FileStream fileStream = new(videoPath, FileMode.Open, FileAccess.Read);
-            await bot.SendVideoAsync(chatId: userId, video: fileStream, replyMarkup: await InlineKeyboard.GetVideoPostKeyboard(userId, videoUrl));
-            fileStream.Close();
-            await bot.DeleteMessageAsync(chatId: userId, messageId: videoIsDownloading_msg.MessageId);
-            await bot.DeleteMessageAsync(chatId: userId, messageId: msg.MessageId);
-            System.IO.File.Delete(videoPath);
-        }
-        public static string LinkCorrector(string link) {
-            try {
-                if (link.Contains("vm.tiktok.com")) { return link; } else { return link.Substring(0, link.IndexOf('?')); }
-            } catch { return link; }
         }
         public static async Task<bool> CanVideoBePostedToChannel (long userId, string channel) {
             bool userCanPostMessages = false;
@@ -65,11 +66,11 @@ namespace SendersNS {
                     if (admin.User.Id == userId) {
                         userCanPostMessages = true;
                     }
-                    if (admin.User.Id == bot.BotId) {
+                    if (admin.User.Id == bot.BotId && (bool)admin.CanPostMessages) {
                         botCanPostMessages = true;
                     }
                 }
-            } catch { Logs.GetErrorLogAboutDeleteBotFromChannel(userId,channelName); return false; }
+            } catch { await Logs.GetErrorLogAboutDeleteBotFromChannelAsync(userId, channelName); return false; }
             return userCanPostMessages&&botCanPostMessages;
         }
     }
